reformatted tg_bot

+++ app/models/user.py
@@ -1,5 +1,5 @@
 from fastapi_users_db_sqlalchemy import SQLAlchemyBaseUserTable
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.db import Base
@@ -7,6 +7,4 @@ from app.core.db import Base
 
 class User(SQLAlchemyBaseUserTable[int], Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    telegram_id: Mapped[int] = mapped_column(
-        Integer, unique=True, nullable=True
-    )
+    telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=True)
